Placeholder for SAI implement
