+++ dataplane/standalone/BUILD.bazel
@@ -4,19 +4,22 @@ cc_library(
         "sai_entrypoint.cc",
         "translator.cc",
     ],
+    alwayslink = True,
     hdrs = ["translator.h"],
     deps = [
+        "//dataplane/standalone/lucius:lucius_clib",
         "//proto/forwarding:forwarding_cc_grpc",
         "//proto/forwarding:forwarding_cc_proto",
         "@com_github_opencomputeproject_sai//:sai",
-        "//dataplane/saigo/standalone",
     ],
 )
 
-cc_shared_library(
-    name = "saigo_shared",
+
+
+cc_binary(
+    name = "sai",
+    linkshared = True,
     deps = [
         ":saigo",
     ],
-    shared_lib_name = "libsai.so",
 )
